+++ project/urls.py
@@ -40,34 +40,44 @@ urlpatterns = [
     ),
     # REPORT
     path(
+        route="<int:pk>/tableau-de-bord/artificialisation",
+        view=views.diagnostic.DiagnosticArtificialisationView.as_view(),
+        name="report_artif",
+    ),
+    path(
+        route="<int:pk>/tableau-de-bord/impermeabilisation",
+        view=views.diagnostic.DiagnosticImpermeabilisationView.as_view(),
+        name="report_imper",
+    ),
+    path(
         "<int:pk>/tableau-de-bord/consommation",
-        views.ProjectReportConsoView.as_view(),
+        views.diagnostic.DiagnosticConsoView.as_view(),
         name="report_conso",
     ),
     path(
         "<int:pk>/tableau-de-bord/synthesis",
-        views.ProjectReportSynthesisView.as_view(),
+        views.diagnostic.DiagnosticConsoView.as_view(),
         name="report_synthesis",
     ),
     path(
         "<int:pk>/tableau-de-bord/vacance-des-logements",
-        views.ProjectReportLogementVacantView.as_view(),
+        views.diagnostic.DiagnosticLogementVacantView.as_view(),
         name="report_logement_vacant",
     ),
     path(
         "<int:pk>/tableau-de-bord/trajectoires",
-        views.ProjectReportTarget2031View.as_view(),
+        views.diagnostic.DiagnosticTarget2031View.as_view(),
         name="report_target_2031",
     ),
     path(
         "<int:pk>/tableau-de-bord/rapport-local",
-        views.ProjectReportLocalView.as_view(),
+        views.diagnostic.DiagnostictRapportLocalView.as_view(),
         name="report_local",
     ),
     # REPORT PARTIALS
     path(
         "<int:pk>/target-2031-graphic",
-        views.ProjectReportTarget2031GraphView.as_view(),
+        views.diagnostic.DiagnosticTarget2031GraphView.as_view(),
         name="target-2031-graphic",
     ),
     # MAP
@@ -79,7 +89,7 @@ urlpatterns = [
     # DOWNLOAD
     path(
         "<int:pk>/tableau-de-bord/telechargement/<slug:requested_document>",
-        views.ProjectReportDownloadView.as_view(),
+        views.diagnostic.DiagnosticDownloadWordView.as_view(),
         name="report_download",
     ),
     path(
@@ -89,11 +99,10 @@ urlpatterns = [
     ),
     path(
         "<int:request_id>/word/telechargement",
-        views.DownloadWordView.as_view(),
+        views.diagnostic.DiagnosticDownloadWordView.as_view(),
         name="word_download",
     ),
 ]
-
 
 # Add API urls
 router = routers.DefaultRouter()
