+++ project/views/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 from .crud import *
 from .dev import *
+from .diagnostic import *
 from .map import *
-from .report import *
