+++ project/views/diagnostic/DiagnosticArtificialisationView.py
@@ -1,0 +1,6 @@
+from .DiagnosticBaseView import DiagnosticBaseView
+
+
+class DiagnosticArtificialisationView(DiagnosticBaseView):
+    partial_template_name = None
+    full_template_name = "project/pages/artificialisation.html"
