+++ project/views/diagnostic/DiagnosticImpermeabilisationView.py
@@ -1,0 +1,6 @@
+from .DiagnosticBaseView import DiagnosticBaseView
+
+
+class DiagnosticImpermeabilisationView(DiagnosticBaseView):
+    partial_template_name = None
+    full_template_name = "project/pages/impermeabilisation.html"
