+++ project/views/diagnostic/__init__.py
@@ -1,0 +1,25 @@
+from .DiagnosticArtificialisationView import DiagnosticArtificialisationView
+from .DiagnosticBaseView import DiagnosticBaseView
+from .DiagnosticConsoView import DiagnosticConsoView
+from .DiagnosticDownloadView import DiagnosticDownloadView
+from .DiagnosticDownloadWordView import DiagnosticDownloadWordView
+from .DiagnosticImpermeabilisationView import DiagnosticImpermeabilisationView
+from .DiagnosticLogementVacantView import DiagnosticLogementVacantView
+from .DiagnosticSynthesisView import DiagnosticSynthesisView
+from .DiagnosticTarget2031GraphView import DiagnosticTarget2031GraphView
+from .DiagnosticTarget2031View import DiagnosticTarget2031View
+from .DiagnostictRapportLocalView import DiagnostictRapportLocalView
+
+__all__ = [
+    "DiagnosticBaseView",
+    "DiagnosticConsoView",
+    "DiagnosticDownloadView",
+    "DiagnosticDownloadWordView",
+    "DiagnosticLogementVacantView",
+    "DiagnosticSynthesisView",
+    "DiagnosticTarget2031GraphView",
+    "DiagnosticTarget2031View",
+    "DiagnostictRapportLocalView",
+    "DiagnosticArtificialisationView",
+    "DiagnosticImpermeabilisationView",
+]
