+++ project/views/mixins.py
@@ -61,6 +61,8 @@ class ReactMixin:
 
     def get_template_names(self):
         if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
+            if not self.partial_template_name:
+                raise ValueError("Partial template name must be defined.")
             return [self.partial_template_name]
         return [self.full_template_name]
